Fixing locking error

+++ arangod/RocksDBEngine/RocksDBIndex.cpp
@@ -197,7 +197,7 @@ void RocksDBIndex::blackListKey(char const* data, std::size_t len){
     bool blacklisted = false;
     uint64_t attempts = 0;
     while (!blacklisted) {
-      blacklisted = _cache->blacklist(data,len);
+      blacklisted = _cache->blacklist(data, (uint32_t)len);
       if (attempts++ % 10 == 0) {
         if (_cache->isShutdown()) {
           disableCache();
